refactor: frameworks/environments/habitat.py -> simulators/habitat/environment.py
